Fix error when running resource actions
