+++ routes/cp.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use StatamicRadPack\Runway\Http\Controllers\CP\ModelRevisionsController;
 use StatamicRadPack\Runway\Http\Controllers\CP\PublishedModelsController;
+use StatamicRadPack\Runway\Http\Controllers\CP\ModelActionController;
 use StatamicRadPack\Runway\Http\Controllers\CP\ResourceActionController;
 use StatamicRadPack\Runway\Http\Controllers\CP\ResourceController;
 use StatamicRadPack\Runway\Http\Controllers\CP\ResourceListingController;
@@ -10,10 +11,13 @@ use StatamicRadPack\Runway\Http\Controllers\CP\RestoreModelRevisionController;
 
 Route::name('runway.')->prefix('runway')->group(function () {
     Route::get('/{resource}', [ResourceController::class, 'index'])->name('index');
-
     Route::get('{resource}/listing-api', [ResourceListingController::class, 'index'])->name('listing-api');
-    Route::post('{resource}/actions', [ResourceActionController::class, 'runAction'])->name('actions.run');
+
+    Route::post('{resource}/actions', [ResourceActionController::class, 'run'])->name('actions.run');
     Route::post('{resource}/actions/list', [ResourceActionController::class, 'bulkActionsList'])->name('actions.bulk');
+
+    Route::post('{resource}/models/actions', [ModelActionController::class, 'runAction'])->name('models.actions.run');
+    Route::post('{resource}/models/actions/list', [ModelActionController::class, 'bulkActionsList'])->name('models.actions.bulk');
 
     Route::get('{resource}/create', [ResourceController::class, 'create'])->name('create');
     Route::post('{resource}/create', [ResourceController::class, 'store'])->name('store');
