+++ src/Fieldtypes/HasManyFieldtype.php
@@ -61,7 +61,7 @@ class HasManyFieldtype extends BaseFieldtype
     public function preload()
     {
         return array_merge(parent::preload(), [
-            'actionUrl' => cp_route('runway.actions.run', [
+            'actionUrl' => cp_route('runway.models.actions.run', [
                 'resource' => $this->config('resource'),
             ]),
         ]);
